Добавила абстрактный компонент
