+++ src/components/button-show-more.js
@@ -1,24 +1,8 @@
-import {createElement} from "../utils";
+import AbstractComponent from "./abstract-component";
 
-class ButtonShowMore {
-  constructor() {
-    this._element = null;
-  }
-
-  get element() {
-    if (this._element === null) {
-      this._element = createElement(this.template);
-    }
-
-    return this._element;
-  }
-
+class ButtonShowMore extends AbstractComponent {
   get template() {
     return `<button class="films-list__show-more">Show more</button>`;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
